+++ assets/docs/Changelog.md
@@ -11,6 +11,7 @@
 ### 其他
 - 移除 power_total、dispatch_count 配置项
 - 使用消耗品前会先筛选类别避免背包物品太多
+- 升级 [PaddleOCR-json_v.1.3.1](https://github.com/hiroi-sora/PaddleOCR-json/releases/tag/v1.3.1)，兼容 Win7 x64
 
 ## v1.6.1